requirements.txt
future==0.16
psycopg2-binary==2.7.7
django==1.11.20
djangorestframework==3.8
djangorestframework-gis==0.13
drf-nested-routers==0.91
django_extensions==2.1.6
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django_filter==1.1
boto3==1.9.126
unicodecsv==0.14.1
gunicorn==19.9.0
supervisor==3.3.5
fabric==2.4.0

simpleflake==0.1.5
rest_condition==1.0.3
python-jose==1.4.0
auth0-python==3.1.2
watchtower==0.3.3
pytest==3.3.2
pytest-django==3.1.2
pytest-cov==2.5.1
pylint==1.8.1
pylint-django==0.8.0
mock==2.0.0
django-taggit==0.23.0
flake8==3.7.7
django-maintenance-mode==0.13.1
mailchimp3==3.0.7
